refactor(docs-site): extract the answer-to-element table module
