+++ apps/docs-site/src/lib/delivery-topology.ts
@@ -1,129 +1,37 @@
 /**
- * The answer-to-element table behind the independence-seam diagram, and the pure
- * derivation that turns an evaluated assessment into the model it draws.
+ * The pure derivation that turns an evaluated assessment into the model the
+ * independence-seam diagram draws.
  *
- * The governing rule is that an element may be drawn only if a named answer id
- * produces it. Every mark therefore appears here as a row of
- * {@link DELIVERY_TOPOLOGY_TABLE}, and an answer absent from the table draws
- * nothing, which is what keeps a delivery diagram from drifting into
- * registries, environments and pipeline stages the assessment never asked about.
+ * Every mark it produces traces to a row of the answer-to-element table, or to
+ * the dataset itself, so an element the reader's answers never established is
+ * drawn as a placeholder rather than guessed at.
  *
  * @module delivery-topology
  */
 
 import type { Answer, Family, Question } from '../data/decision-framework'
 import type { EngineResult } from './decision-engine'
+import type {
+  Admission,
+  BoundaryLanguage,
+  ControlGlyph,
+  DividerMark,
+  PageShape,
+  PartyLane,
+  RegionLink,
+  Seat,
+  SeamState,
+  TopologyRule,
+} from './delivery-topology-table'
 import { createMap } from '@hyperfrontend/immutable-api-utils/built-in-copy/map'
 import { decisionFramework } from '../data/decision-framework'
 import { HYPERFRONTEND_FAMILY_ID } from './decision-engine'
+import { ruleFor, YOUR_TEAM } from './delivery-topology-table'
 
-/** How a party lane is filled. One treatment per epistemic state, never shared. */
-export type PartyTreatment = 'your-team' | 'another-team' | 'cannot-direct' | 'not-established'
 /** Delivery-rail treatment, encoding how strongly the answer behind a band binds. */
 export type RailStyle = 'hard' | 'strong-preference' | 'weak-preference' | 'unanswered'
-/** Which directions of the ownership arrow carry a slash. */
-export type ControlGlyph = 'none' | 'one-way' | 'both-ways'
-/** The four states of the seam, the one mark that carries the verdict. */
-export type SeamState = 'welded' | 'gapped-hard' | 'gapped-by-preference' | 'undecided'
-/** The only two vocabularies the boundary may use. Adversarial wording is earned, never assumed. */
-export type BoundaryLanguage = 'accidents' | 'attackers'
-/** How a new piece reaches the running page. */
-export type AdmissionState = 'runtime-admission' | 'redeploy-per-piece'
-/** What the user sees: one screen carrying two owners, or one owner per page. */
-export type PageShape = 'mixed-screen' | 'mixed-screen-later' | 'page-per-team' | 'undetermined'
-/** What divides one owner's region from another. */
-export type DividerGlyph = 'seam-line' | 'wall' | 'double-wall'
-/** Which seat the reader occupies in the composition. */
-export type SeatState = 'reader-hosts' | 'reader-participates' | 'varies-per-host'
 /** The assembly-locus poles, which the dataset separates into three rather than two. */
 export type CompositionPole = 'build' | 'request-path' | 'browser'
-/** How two regions talk once they share a screen. */
-export type RegionLinkKind = 'same-call-stack' | 'messages'
-
-/** One party in band 1: who builds and ships a piece. */
-export interface PartyLane {
-  /** Stable lane id, used as the render key. */
-  id: string
-  /** The lane label in words, so no meaning rests on fill or hue. */
-  label: string
-  /** How the lane is filled, which is a second channel over the label rather than a substitute for it. */
-  treatment: PartyTreatment
-}
-
-/** The mark that divides regions in band 4, with the wording that earns it. */
-export interface DividerMark {
-  /** Which of the three glyphs is drawn. */
-  glyph: DividerGlyph
-  /** What the reader stated, phrased as a requirement rather than as a delivered barrier. */
-  label: string
-}
-/** How the parts talk across the boundary once they share a screen. */
-export interface RegionLink {
-  /** Whether one call stack is required or messages are acceptable. */
-  kind: RegionLinkKind
-  /** The reader's own constraint, in one phrase. */
-  text: string
-}
-/** Which seat the reader occupies, and the answer that established it. */
-export interface Seat {
-  /** Whether the reader hosts, participates, or cannot say. */
-  state: SeatState
-  /** The circumstance the reader reported, in one phrase. */
-  note: string
-}
-/** How a new piece is admitted into the running product. */
-export interface Admission {
-  /** Whether admission happens at runtime or through a redeploy. */
-  state: AdmissionState
-  /** The requirement in one phrase. */
-  text: string
-}
-
-/**
- * One row of the answer-to-element table: the only licence any mark has to exist.
- *
- * Every field is optional because one answer produces one or two elements and
- * nothing else. An answer that produces no element has no row, which is the
- * table working rather than a gap in it.
- */
-export interface TopologyRule {
-  /** The answer id that licenses these elements. */
-  answerId: string
-  /** Party lanes established in band 1. */
-  lanes?: readonly PartyLane[]
-  /** Whether band 1 draws the ghost strip standing for parties the assessment never counts. */
-  ghostStrip?: boolean
-  /** The control glyph in band 1. */
-  controlGlyph?: ControlGlyph
-  /** What the control glyph asserts, in words. */
-  controlNote?: string
-  /** What every party hands over, in band 2. */
-  handover?: string
-  /** Whether the handover carries the locked glyph, for a piece nothing may change. */
-  handoverLocked?: boolean
-  /** The seam state in band 3. */
-  seam?: SeamState
-  /** Whether the seam follows from an ownership fact rather than from a choice. */
-  seamEntailed?: boolean
-  /** Who runs the composition point, in band 3. */
-  operator?: string
-  /** The vocabulary the boundary may use. */
-  boundaryLanguage?: BoundaryLanguage
-  /** A requirement the reader stated, rendered as a chip that never promises delivery. */
-  requirement?: string
-  /** How new pieces join the running page, in band 3. */
-  admission?: Admission
-  /** The page shape in band 4. */
-  pageShape?: PageShape
-  /** The divider between regions in band 4. */
-  divider?: DividerMark
-  /** How the regions talk across the divider. */
-  regionLink?: RegionLink
-  /** A framework-copy annotation, which only a mixed screen ever raises. */
-  dependencyNote?: string
-  /** Which seat the reader occupies. */
-  seat?: Seat
-}
 
 /** Short authored copy for an element no answer has filled in yet. */
 export interface PlaceholderCopy {
@@ -343,134 +251,6 @@ export interface DeliveryTopology {
   announcement: string
 }
 
-const YOUR_TEAM: PartyLane = { id: 'party.you', label: 'your team', treatment: 'your-team' }
-const ANOTHER_TEAM: PartyLane = { id: 'party.another', label: 'another team at your company', treatment: 'another-team' }
-const OUTSIDE_PARTY: PartyLane = { id: 'party.outside', label: 'a party you cannot direct', treatment: 'cannot-direct' }
-
-/** The answer-to-element table: a mark with no row here is a mark with no evidence. */
-export const DELIVERY_TOPOLOGY_TABLE: readonly TopologyRule[] = [
-  { answerId: 'question.ownership.composition-parties#one-team', lanes: [YOUR_TEAM] },
-  { answerId: 'question.ownership.composition-parties#several-teams', lanes: [YOUR_TEAM, ANOTHER_TEAM], ghostStrip: true },
-  {
-    answerId: 'question.ownership.composition-parties#outside-party',
-    lanes: [YOUR_TEAM, OUTSIDE_PARTY],
-    ghostStrip: true,
-    controlGlyph: 'one-way',
-    controlNote: 'you cannot direct when they ship',
-  },
-  {
-    answerId: 'question.ownership.composition-parties#no-deploy-control',
-    lanes: [YOUR_TEAM, OUTSIDE_PARTY],
-    ghostStrip: true,
-    controlGlyph: 'both-ways',
-    controlNote: 'neither side can direct when the other ships',
-    seam: 'gapped-hard',
-    seamEntailed: true,
-  },
-  { answerId: 'question.deploy.independence#independent', seam: 'gapped-hard' },
-  { answerId: 'question.deploy.independence#valuable-not-required', seam: 'gapped-by-preference' },
-  { answerId: 'question.deploy.independence#train-mandated', seam: 'welded' },
-  {
-    answerId: 'question.migration.participant-ceiling#refactor-into-codebase',
-    handover: 'source that can be refactored into a shared codebase',
-  },
-  { answerId: 'question.migration.participant-ceiling#bootstrap-edit', handover: 'a separate app whose build and startup can change' },
-  {
-    answerId: 'question.migration.participant-ceiling#build-change-only',
-    handover: 'a separate app whose build can change, but not its startup',
-  },
-  {
-    answerId: 'question.migration.participant-ceiling#wrap-as-is',
-    handover: 'a deployed app, exactly as it already runs',
-    handoverLocked: true,
-  },
-  {
-    answerId: 'question.delivery.server-capacity#static-only',
-    operator: 'you ship files to a CDN and operate nothing on the request path',
-  },
-  { answerId: 'question.delivery.server-capacity#no-new-tier', operator: 'servers exist, but nobody wants another service to operate' },
-  {
-    answerId: 'question.delivery.server-capacity#operates-servers',
-    operator: 'a team already runs server estates and could take one more',
-  },
-  {
-    answerId: 'question.trust.malicious-participant#contain-malice',
-    boundaryLanguage: 'attackers',
-    requirement: 'you required: it holds even if that piece is compromised',
-    divider: { glyph: 'double-wall', label: 'a doubled wall: it holds even if that piece is compromised' },
-  },
-  {
-    answerId: 'question.failure.containment#must-survive',
-    requirement: 'you required: one region may fail without the others',
-    divider: { glyph: 'wall', label: 'a wall: one region may fail without the others' },
-  },
-  {
-    answerId: 'question.ux.seam-tolerance#seams-acceptable',
-    divider: { glyph: 'seam-line', label: 'a visible seam: edges are acceptable here' },
-  },
-  {
-    answerId: 'question.delivery.first-paint#crawlable-required',
-    requirement: 'you required: the combined pages render before any script runs',
-  },
-  {
-    answerId: 'question.roster.runtime-admission#no-host-change',
-    admission: { state: 'runtime-admission', text: 'new pieces and new versions appear without anyone touching the host' },
-  },
-  {
-    answerId: 'question.roster.runtime-admission#redeploy-acceptable',
-    admission: { state: 'redeploy-per-piece', text: 'each new piece is admitted by redeploying the main application' },
-  },
-  { answerId: 'question.granularity.single-screen#mixed-screen', pageShape: 'mixed-screen' },
-  { answerId: 'question.granularity.single-screen#probably-later', pageShape: 'mixed-screen-later' },
-  { answerId: 'question.granularity.single-screen#page-per-team', pageShape: 'page-per-team' },
-  {
-    answerId: 'question.contracts.sync-calls#sync-required',
-    regionLink: { kind: 'same-call-stack', text: 'one live call stack is required' },
-  },
-  { answerId: 'question.contracts.sync-calls#messaging-acceptable', regionLink: { kind: 'messages', text: 'messages cross the boundary' } },
-  {
-    answerId: 'question.coordination.upgrade-train#skew-today',
-    dependencyNote: 'each team upgrades when it can, and the screen carries the skew',
-  },
-  {
-    answerId: 'question.coordination.upgrade-train#could-govern',
-    dependencyNote: 'teams could move in step, but you would rather not depend on it',
-  },
-  {
-    answerId: 'question.coordination.upgrade-train#trains-run',
-    dependencyNote: 'aligned upgrade trains already run across the affected teams',
-  },
-  { answerId: 'question.deps.major-coexistence#stuck-and-unfunded', dependencyNote: 'some pieces are stuck on an older major, unfunded' },
-  {
-    answerId: 'question.deps.major-coexistence#alignment-funded',
-    dependencyNote: 'versions are mixed, and the alignment work is funded and owned',
-  },
-  {
-    answerId: 'question.deps.major-coexistence#aligned',
-    dependencyNote: 'everything already runs the same major, or the differences are compatible',
-  },
-  {
-    answerId: 'question.deps.payload-budget#hard-budget',
-    dependencyNote: 'one page-weight budget applies to every piece on the screen together',
-  },
-  { answerId: 'question.deps.payload-budget#duplication-tolerable', dependencyNote: 'some duplication between the pieces is tolerable' },
-  {
-    answerId: 'question.host.negotiability#hosts-unmodifiable',
-    seat: {
-      state: 'reader-participates',
-      note: 'hosts cannot be asked to change anything, so your product runs in their pages as they are',
-    },
-  },
-  {
-    answerId: 'question.host.negotiability#credible-ask',
-    seat: { state: 'varies-per-host', note: 'most hosts would adopt a small install, but not all of them will' },
-  },
-  {
-    answerId: 'question.host.negotiability#hosts-cooperate',
-    seat: { state: 'reader-hosts', note: 'every embedding host is yours, or cooperates' },
-  },
-]
-
 /** Authored short copy for the unanswered state, one entry per question that draws a placeholder. */
 const PLACEHOLDER_COPY: Record<string, PlaceholderCopy> = {
   'question.ownership.composition-parties': { prompt: 'Who else ships a piece? Not asked yet', subject: 'who builds and ships each piece' },
@@ -540,8 +320,6 @@ const BAND_QUESTIONS: Record<string, readonly string[]> = {
 
 const SEAT_UNANSWERED: Seat = { state: 'reader-hosts', note: 'nothing you answered moves the host seat, so this figure keeps you in it' }
 
-const RULE_INDEX = createMap(DELIVERY_TOPOLOGY_TABLE.map((rule) => <readonly [string, TopologyRule]>[rule.answerId, rule]))
-
 const questionEntries: [string, Question][] = []
 const answerEntries: [string, Answer][] = []
 for (const question of decisionFramework.questions) {
@@ -550,16 +328,6 @@ for (const question of decisionFramework.questions) {
 }
 const QUESTION_INDEX = createMap(questionEntries)
 const ANSWER_INDEX = createMap(answerEntries)
-
-/**
- * Looks up the row that licenses an answer to draw something.
- *
- * @param answerId - The answer the reader selected.
- * @returns The row, or undefined when the answer draws nothing at all.
- */
-export function ruleFor(answerId: string): TopologyRule | undefined {
-  return RULE_INDEX.get(answerId)
-}
 
 const lastOf = <T>(rules: readonly TopologyRule[], read: (rule: TopologyRule) => T | undefined): T | undefined => {
   let found: T | undefined
@@ -753,7 +521,7 @@ const thesisOf = (topology: Omit<DeliveryTopology, 'thesis' | 'announcement'>): 
 /**
  * Turns an evaluated assessment into the topology the independence-seam diagram
  * draws. Pure: the same result always produces the same model, and every mark in
- * it traces to a row of {@link DELIVERY_TOPOLOGY_TABLE} or to the dataset itself.
+ * it traces to a row of the answer-to-element table or to the dataset itself.
  *
  * @param result - The evaluated assessment.
  * @returns The four bands, the thesis, and the live-region sentence.
